backend/app/nft_transactions/requirements.txt
anyio==3.3.4
base58==2.1.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
ECPy==1.2.5
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
python-dotenv==0.19.2
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.26
SQLAlchemy-Utils==0.38.2
typing-extensions==3.10.0.2
urllib3==1.26.7
websockets==9.1
Werkzeug==2.0.2
xrpl-py==1.3.0
